FF do develop branch

+++ deployment/puppet/quantum/manifests/agents/ovs.pp
@@ -16,6 +16,19 @@ class quantum::agents::ovs (
 
   include 'quantum::waist_setup'
 
+  if defined(Anchor['quantum-plugin-ovs-done']) {
+    # install quantum-ovs-agent at the same host where 
+    # quantum-server + quantum-ovs-plugin
+    Anchor['quantum-plugin-ovs-done'] -> Anchor['quantum-ovs-agent']
+
+  }
+
+  if defined(Anchor['quantum-server-done']) {
+    Anchor['quantum-server-done'] -> Anchor['quantum-ovs-agent']
+  }
+
+  anchor {'quantum-ovs-agent': }
+
   if $::quantum::params::ovs_agent_package {
     Package['quantum'] -> Package['quantum-plugin-ovs-agent']
 
@@ -29,7 +42,10 @@ class quantum::agents::ovs (
     $ovs_agent_package = $::quantum::params::ovs_server_package
   }
 
-  Package[$ovs_agent_package] -> Quantum_plugin_ovs <| |>
+  if !defined(Anchor['quantum-server-done']) {
+    # if defined -- this depends already defined
+    Package[$ovs_agent_package] -> Quantum_plugin_ovs <| |>
+  }
 
   l23network::l2::bridge { $integration_bridge:
     external_ids  => "bridge-id=${integration_bridge}",
@@ -38,17 +54,24 @@ class quantum::agents::ovs (
   }
 
   if $enable_tunneling {
+    L23network::L2::Bridge<| |> ->
+      Anchor['quantum-ovs-agent-done']
     l23network::l2::bridge { $tunnel_bridge:
       external_ids  => "bridge-id=${tunnel_bridge}",
       ensure        => present,
       skip_existing => true,
-    }
+    } ->
+    Anchor['quantum-ovs-agent-done']
     quantum_plugin_ovs { 'OVS/local_ip': value => $local_ip; }
+
   } else {
+    L23network::L2::Bridge[$integration_bridge] ->
+      Anchor['quantum-ovs-agent-done']
     quantum::plugins::ovs::bridge { $bridge_mappings: # Do not quote!!! may be array!
-    }
+    } ->
     quantum::plugins::ovs::port { $bridge_uplinks: # Do not quote!!! may be array!
-    }
+    } -> 
+    Anchor['quantum-ovs-agent-done']
   }
 
   if $enabled {
@@ -57,22 +80,17 @@ class quantum::agents::ovs (
     $service_ensure = 'stopped'
   }
 
-  Quantum_config <| |> ~> Service['quantum-ovs-agent']
-  Quantum_plugin_ovs <| |> ~> Service['quantum-ovs-agent']
-  Service <| title == 'quantum-server' |> -> Service['quantum-ovs-agent']
+  #Quantum_config <| |> ~> Service['quantum-ovs-agent']
+  #Quantum_plugin_ovs <| |> ~> Service['quantum-ovs-agent']
+  #Service <| title == 'quantum-server' |> -> Service['quantum-ovs-agent']
 
-  L23network::L2::Bridge <| |> -> Service['quantum-ovs-agent']
-
-  if $service_provider == 'pacemaker' {
+    if $service_provider == 'pacemaker' {
     Quantum_config <| |> -> Cs_shadow['ovs']
     Quantum_plugin_ovs <| |> -> Cs_shadow['ovs']
     L23network::L2::Bridge <| |> -> Cs_shadow['ovs']
 
     cs_shadow { 'ovs': cib => 'ovs' }
-
     cs_commit { 'ovs': cib => 'ovs' }
-
-    Cs_commit['ovs'] -> Service['quantum-ovs-agent']
 
     ::corosync::cleanup { "p_${::quantum::params::ovs_agent_service}": }
 
@@ -80,6 +98,7 @@ class quantum::agents::ovs (
     Cs_commit['ovs'] ~> ::Corosync::Cleanup["p_${::quantum::params::ovs_agent_service}"]
     ::Corosync::Cleanup["p_${::quantum::params::ovs_agent_service}"] -> Service['quantum-ovs-agent']
 
+    File<| title=='quantum-logging.conf' |> ->
     cs_resource { "p_${::quantum::params::ovs_agent_service}":
       ensure          => present,
       cib             => 'ovs',
@@ -126,18 +145,24 @@ class quantum::agents::ovs (
       name       => $::quantum::params::ovs_agent_service,
       enable     => false,
       hasstatus  => false,
+      hasrestart => false
     }
-    exec { 'quantum-ovs-agent_stopped':
-      #todo: rewrite as script, that returns zero or wait, when it can return zero
-      name   => "bash -c \"service ${::quantum::params::ovs_agent_service} stop || ( kill `pgrep -f quantum-openvswitch-agent` || : )\"",
-      onlyif => "service ${::quantum::params::ovs_agent_service} status | grep \'${started_status}\'",
-      path   => ['/usr/bin', '/usr/sbin', '/bin', '/sbin'],
-      returns => [0,""]
+
+    if $::osfamily =~ /(?i)debian/ {
+      exec { 'quantum-ovs-agent_stopped':
+        #todo: rewrite as script, that returns zero or wait, when it can return zero
+        name   => "bash -c \"service ${::quantum::params::ovs_agent_service} stop || ( kill `pgrep -f quantum-openvswitch-agent` || : )\"",
+        onlyif => "service ${::quantum::params::ovs_agent_service} status | grep \'${started_status}\'",
+        path   => ['/usr/bin', '/usr/sbin', '/bin', '/sbin'],
+        returns => [0,""]
+      }
     }
-    Package[$ovs_agent_package] ->
-      Service['quantum-ovs-agent_stopped'] ->
-        Exec['quantum-ovs-agent_stopped'] ->
-          Cs_resource["p_${::quantum::params::ovs_agent_service}"]
+    L23network::L2::Bridge<| |> -> 
+      Package[$ovs_agent_package] ->
+        Service['quantum-ovs-agent_stopped'] ->
+          Exec<| title=='quantum-ovs-agent_stopped' |> ->
+            Cs_resource["p_${::quantum::params::ovs_agent_service}"] -> 
+              Service['quantum-ovs-agent']
 
     service { 'quantum-ovs-agent':
       name       => "p_${::quantum::params::ovs_agent_service}",
@@ -158,9 +183,10 @@ class quantum::agents::ovs (
       provider   => $::quantum::params::service_provider,
     }
   }
-  Class[quantum::waistline] -> Service['quantum-ovs-agent']
-  Package[$ovs_agent_package] -> Service['quantum-ovs-agent']
 
+  Class[quantum::waistline] -> Service['quantum-ovs-agent']
+
+  #todo: This service must be disabled if Quantum-ovs-agent managed by pacemaker
   service { 'quantum-ovs-cleanup':
     name       => 'quantum-ovs-cleanup',
     enable     => $enabled,
@@ -169,8 +195,16 @@ class quantum::agents::ovs (
     hasrestart => false,  # !!! cleanup is simple script runnung once at OS boot
   }
 
-  # This order is conditioned package dependencies
-  # In real life, the starting order of the inverse
-  Service['quantum-ovs-agent'] -> Service['quantum-ovs-cleanup']
+  Anchor['quantum-ovs-agent'] -> 
+    Service['quantum-ovs-agent'] ->       # it's not mistate!
+      Service['quantum-ovs-cleanup'] ->   # cleanup service after agent.
+        Anchor['quantum-ovs-agent-done']
+
+  anchor{'quantum-ovs-agent-done': }
+
+  Anchor['quantum-ovs-agent-done'] -> Anchor<| title=='quantum-l3' |>
+  Anchor['quantum-ovs-agent-done'] -> Anchor<| title=='quantum-dhcp-agent' |>
 
 }
+#
+###
